バグを修正(#3)::パースエラー(プロガード対応)
