+++ app/src/main/kotlin/jp/co/yumemi/android/codeCheck/data/models/GitHubRepository.kt
@@ -1,0 +1,15 @@
+package jp.co.yumemi.android.codeCheck.data.models
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class GitHubRepository(
+    val name: String,
+    val ownerIconUrl: String,
+    val language: String,
+    val stargazersCount: Long,
+    val watchersCount: Long,
+    val forksCount: Long,
+    val openIssuesCount: Long,
+) : Parcelable
