+++ app/src/main/res/navigation/nav_graph.xml
@@ -21,6 +21,6 @@
         tools:layout="@layout/fragment_detail_repository" >
         <argument
             android:name="item"
-            app:argType="jp.co.yumemi.android.codeCheck.ui.GitHubRepository" />
+            app:argType="jp.co.yumemi.android.codeCheck.data.models.GitHubRepository" />
     </fragment>
 </navigation>
